Fixed a Visual Studio warning

+++ source/r_interpolate.h
@@ -57,7 +57,7 @@ inline fixed_t lerpCoord(fixed_t lerp, fixed_t oldpos, fixed_t newpos)
 // Intentionally give a different name to prevent accidental overloads
 inline float lerpCoordf(fixed_t lerp, float oldpos, float newpos)
 {
-   return oldpos + lerp / FPFRACUNIT * (newpos - oldpos);
+   return oldpos + lerp / (float)FPFRACUNIT * (newpos - oldpos);
 }
 
 #define LLANG360 4294967296LL
